Don't store nKeyLength for integer and interned keys
Previously hash buckets always stored an nKeyLength member, even
for integer keys and keys using interned strings. In both cases
this information is not necessary, because integers have no length
and interned strings already store their own length separately.

The new behavior is to store nKeyLength only for non-interned string
keys. The value is stored immidiately before the actual string data.
As such the bucket layout is now as follows:

    bucket  sizeof(Bucket)
    data    ZEND_MM_ALIGNED_SIZE(nDataSize)
   [key_len ZEND_MM_ALIGNED_SIZE(zend_uint)]
   [key_str nKeyLength                     ]

This saves 4 bytes (32 bit) or 8 bytes (64 bit) for every bucket,
which has an integer or interned string key.

API changes:

 * Code making direct use of the Bucket structure must use
   p->arKey != NULL to detect string keys instead of
   p->nKeyLength != NULL.
 * Code making direct use of the Bucket structure must use
   zend_bucket_key_length(p) instead of p->nKeyLength

WARNING: This is an incomplete changeset. It lacks zts and opcache
support.