Implement robin hood hashing
This is a failed attempt at implementing Robin Hood hashing (open
addressing with linear probing that moves elements to minimize
distance to initial bucket). The issue is that RH hashing requires
a high-quality hashing function, which is not the case in PHP. It
is relatively easy to happen on pathological cases without having
to specially engineer them. E.g. hash1() in Zend/bench.php creates
this array:

  for ($i = 1; $i <= $n; $i++) {
    $X[dechex($i)] = $i;
  }

Due to the specifics of the used hash function and how string and
integer keys interact, this will result in the vast majority of
the hashes to be small (after masking). As such nearly all elements
will cluster at the start of arHash (with correspondingly high DIB
values), while the the rest of arHash (with the used allocation
strategy this may be as much as 3/4) is mostly empty.

The original motivation behind using RH hashing was the fact that
is can be used with backshift deletion instead of tombstones (it
actually must be used that way, otherwise deletions will quickly
cause catastrophic DIB values). Not using tombstones means that
we do not need to separate manage utilization of arHash and can
keep arData and arHash in one allocation (if we're okay with using
low load factors of <= 1/2, which I am for now). Splitting them
in separate allocations has in my previous experiments resulted
in very bad cache utilization (an order of magnitude more cache
misses than previously in some cases). It bears further investigation
if this can be avoided.