Remove dependency of oplines on live ranges
This removes the live range number stored in op2 of FREE_ON_RETURN
FREEs/FE_FREEs. Instead we look up the live range at runtime. As
this is only relevant in very unlikely cases, the runtime lookup
is unproblematic.

This simplifies hangling of live ranges in the compiler and
optimizer, as we don't have to scan the oplines after moving
live ranges.